Remove file option and unify the php and abstract process
