--- src/Util/PHP/eval-stdin.php
@@ -1,10 +1,0 @@
-<?php
-/*
- * This file is part of PHPUnit.
- *
- * (c) Sebastian Bergmann <sebastian@phpunit.de>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-eval('?>' . file_get_contents('php://stdin'));
